requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.9
django-ckeditor==6.2.0
django-ckeditor-updated==4.4.4
django-cors-headers==3.10.0
django-filter==21.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.3
mysqlclient==2.1.0
Pillow==8.4.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8